Renaming classes, namespace.
