+++ GoogleCloudExtension/GoogleCloudExtension/PublishDialogSteps/FlexStep/FlexStepViewModel.cs
@@ -17,7 +17,7 @@ using GoogleCloudExtension.Analytics;
 using GoogleCloudExtension.Analytics.Events;
 using GoogleCloudExtension.Deployment;
 using GoogleCloudExtension.GCloud;
-using GoogleCloudExtension.HostAbstraction;
+using GoogleCloudExtension.VsVersion;
 using GoogleCloudExtension.PublishDialog;
 using GoogleCloudExtension.Utils;
 using System;
@@ -133,7 +133,7 @@ namespace GoogleCloudExtension.PublishDialogSteps.FlexStep
                         project,
                         options,
                         progress,
-                        HostAbstractionLayer.ToolsPathProvider,
+                        VsVersionUtils.ToolsPathProvider,
                         GcpOutputWindow.OutputLine);
                     deploymentDuration = DateTime.Now - startDeploymentTime;
                 }
