+++ GoogleCloudExtension/GoogleCloudExtension/PublishDialogSteps/GkeStep/GkeStepViewModel.cs
@@ -19,7 +19,7 @@ using GoogleCloudExtension.Analytics.Events;
 using GoogleCloudExtension.DataSources;
 using GoogleCloudExtension.Deployment;
 using GoogleCloudExtension.GCloud;
-using GoogleCloudExtension.HostAbstraction;
+using GoogleCloudExtension.VsVersion;
 using GoogleCloudExtension.PublishDialog;
 using GoogleCloudExtension.Utils;
 using System;
@@ -273,7 +273,7 @@ namespace GoogleCloudExtension.PublishDialogSteps.GkeStep
                             project,
                             options,
                             progress,
-                            HostAbstractionLayer.ToolsPathProvider,
+                            VsVersionUtils.ToolsPathProvider,
                             GcpOutputWindow.OutputLine);
                         deploymentDuration = DateTime.Now - deploymentStartTime;
                     }
