+++ GoogleCloudExtension/GoogleCloudExtension/VsVersion/VS15/ToolsPathProvider.cs
@@ -17,19 +17,18 @@ using GoogleCloudExtension.Utils;
 using System;
 using System.IO;
 
-namespace GoogleCloudExtension.HostAbstraction.VS14
+namespace GoogleCloudExtension.VsVersion.VS15
 {
     /// <summary>
-    /// The implementation of <seealso cref="IToolsPathProvider"/> for Visual Studio 2015 (v14.0).
+    /// The implementation of <seealso cref="IToolsPathProvider"/> for Visual Studio 2017 (v 15.0).
     /// </summary>
     internal class ToolsPathProvider : IToolsPathProvider
     {
-        public string GetExternalToolsPath()
+        private readonly string _edition;
+
+        public ToolsPathProvider(string edition)
         {
-            var programFilesPath = Environment.GetFolderPath(Environment.SpecialFolder.ProgramFilesX86);
-            var result = Path.Combine(programFilesPath, @"Microsoft Visual Studio 14.0\Web\External");
-            GcpOutputWindow.OutputDebugLine($"External tools path: {result}");
-            return result;
+            _edition = edition;
         }
 
         public string GetDotnetPath()
@@ -40,19 +39,26 @@ namespace GoogleCloudExtension.HostAbstraction.VS14
             return result;
         }
 
+        public string GetExternalToolsPath()
+        {
+            return "";
+        }
+
         public string GetMsbuildPath()
         {
-            var programFilesPath = Environment.GetFolderPath(Environment.SpecialFolder.ProgramFilesX86);
-            var result = Path.Combine(programFilesPath, @"MSBuild\14.0\Bin\MSBuild.exe");
-            GcpOutputWindow.OutputDebugLine($"Msbuild path: {result}");
+            var programFilesPath = Environment.GetEnvironmentVariable("ProgramFiles(x86)");
+            var result = Path.Combine(programFilesPath, $@"Microsoft Visual Studio\2017\{_edition}\MSBuild\15.0\Bin\MSBuild.exe");
+            GcpOutputWindow.OutputDebugLine($"Program Files: {programFilesPath}");
+            GcpOutputWindow.OutputDebugLine($"Msbuild V15 Path: {result}");
             return result;
         }
 
         public string GetMsdeployPath()
         {
-            var programFilesPath = Environment.GetFolderPath(Environment.SpecialFolder.ProgramFiles);
+            var programFilesPath = Environment.GetEnvironmentVariable("ProgramFiles");
             var result = Path.Combine(programFilesPath, @"IIS\Microsoft Web Deploy V3\msdeploy.exe");
-            GcpOutputWindow.OutputDebugLine($"Msdeploy path: {result}");
+            GcpOutputWindow.OutputDebugLine($"Program Files: {programFilesPath}");
+            GcpOutputWindow.OutputDebugLine($"Msdeploy V15 path: {result}");
             return result;
         }
     }
